Moved Auth related actions to authActions
